requirements.txt
Flask>=3.0.0
Flask-RESTX>=1.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
flask_cors